Refactor localization system
This changeset replaces the i18n module with a `dnd.lingo` macro.

The macro emulates the core language system. It checks the value of the
page-level variable `languageTitle`, then searches for a localized
string for that language. Failing that, it searches for the string in
languages marked as "dependents" in the core language plugin. Finally,
it returns a default language string.